Some imporvements while expeirence with TMS320 platform (TI CCS)
* Some markers "DEFINED_... " because a simple platform does not need all features of emC/Base, less emC files.
* StringBase_emC.c: toString_int32_emC(...) without ThCxt, unnecessary, the only one routine in StringBase_emC.c which uses the STACKTRC_ENTRY.
* StringBase_emC.c: prevent some warnings on 64-bit-platform.
* StringBase_emC.h, c is applicable without further dependencies, simple applicable, need for Test/testAssert.*
* StingBase_emC.h is new, contains only independent parts of the String_emC.h
* String_emC.h includes now StringBase_emC.h, it defines more complex string handling.
* emC_inclComplSpec/cc_Msc15_32bit/compl_adaption.h, todo add to all compl_adaption: macro RAMFUNC_emC as function attribute especially for embedded platform.
* ObjectJc_full.h now removed, all included in emC/Base/Object_emC.h

+++ emC/Base/String_emC.c
@@ -87,7 +87,7 @@ int parseIntRadix_emC(const char* srcP, int size, int radix, int* parsedChars)
   }
   if(bNegativ){ val = -val; }
   if(parsedChars !=null){
-		*parsedChars = src - srcP;
+		*parsedChars = (int)(src - srcP);
 	}
   return( val);
 }
